requirements.txt
aiohttp==3.8.3
Cerberus==1.3.4
fastapi==0.85.0
Jinja2==3.1.2
pydantic==1.10.2
python-dateutil==2.8.2
PyYAML==6.0
uvicorn==0.18.3
uvloop==0.17.0
